Initial implementation of SRV lookups.
* src/resolver.c (dns_response_free): Handle dns_srv.
(dns_generic_lookup): New function.
(dns_lookup): Rewrite.
(dns_srv_lookup): New function.
* src/resolver.h (dns_srv): Use int instead of int16_t

+++ src/resolver.h
@@ -42,9 +42,9 @@ union dns_addr
 
 struct dns_srv
 {
-  uint16_t priority;
-  uint16_t weight;
-  uint16_t port;
+  int priority;
+  int weight;
+  int port;
   char *host;
 };
